requirements.txt
langchain
langchain_community
langchain_core
langchain_huggingface
faiss-cpu
python-dotenv
pypdf
huggingface-hub
flask
sentence-transformers
markupsafe


